requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
chromadb
langchain
langchain-google-genai
pypdf
python-multipart
requests
pydantic
python-dotenv
streamlit-mic-recorder
gTTS
SpeechRecognition
